nwag-uw/requirements.txt
dash==2.14.2
dash-bootstrap-components==1.5.0
flask==2.3.3
databricks-sdk>=0.35.0
databricks-sql-connector==2.9.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml==6.0.1
